fix: log level and node organizer
